Added tests for rollback functionality.
To test for rollbacks, we generate some minimal blockchain history with
a mocked out API to sync to for getting blocks to most accurately
simulate an actual sync/rollback. After this we generate an alternate
history and test that the next sync properly rolls back the changes.

These tests are quite computationally expensive due to the fact
that constants 2160/21600 (stability depth/slots per epoch) are
hard-coded in rust/cardano a lot, we have to generate epochs this big.

Also, lots of assumptions in rust-cardano about the epoch you are
syncing to causes syncs to a chain within the first epoch to crash
exacerbate this.

Due to this the test are marked as #[ignore] so they are not triggered
by a quick `cargo test` call. To call them, run `cargo test -- --ignored`,
optionally specifying one or more by name.

The tests do not yet cover the 2nd case when we need to sync far into
the past, but will be updated with that soon.