Retry upon net error during sync (experimental)
Due to the bridge periodically receiving a specific networking-relating
error, this PR attempts to resolve this by restarting up to a threshold
of time when encountered. Due to the nature of the error occuring from
the native (haskell) peers and being networking-related (NTT), it is
unsure if it is a side effect of a larger networking issue. If this is
the case and a total disconnect happens, then the restart will need to
be done at a higher level inside the bridge itself.