Optimize net_sync() - remove unnecessary chain_state restoring.
In order to remove excessive and unnecessary IO work recreating the
`chain_state` when repeatedly syncing when up to date, it was refactored
to be kept in memory instead of repeatedly recreated via
`restore_chain_state()`.

For initial syncs this has basically no performance difference, but when
up to date `net_sync_to()` is called every 20 seconds and this can make
a big difference, as the only other work done is retrieving a single new
block which completely pales in comparison to `restore_chain_state()` -
especially when the chain has many epochs.